Add layer API. Create client/windshaft
