+++ src/api/layer.js
@@ -1,25 +1,67 @@
 import getMGLIntegrator from './mglintegrator';
 
+import * as _ from 'lodash';
+
+import SourceBase from './source/base';
+
 /**
  * Responsabilities:  rely style changes into MNS source notifications, notify renderer about style changes, notify source about viewport changes,
  * rely dataframes to renderer, configure visibility for all source dataframes, set up MGL integration (opionally)
  */
 export default class Layer {
-    constructor(name, source, style) {
-        this._name = name;
-        this._source = source;
-        this._style = style;
+
+    constructor(id, source, style) {
+        this._checkId(id);
+        this._checkSource(source);
+        this._checkStyle(style);
+
         this._lastViewport = null;
         this._lastMNS = null;
         this._mglIntegrator = null;
         this._dataframes = [];
 
-        style.onChange(this._styleChanged.bind(this));
+        this._id = id;
         this.setSource(source);
+        this.setStyle(style);
+
         console.log('L', this);
     }
+
+    _checkId(id) {
+        if (_.isUndefined(id)) {
+            throw new Error('layer', 'idRequired');
+        }
+        if (!_.isString(id)) {
+            throw new Error('layer', 'idString');
+        }
+        if (_.isEmpty(id)) {
+            throw new Error('layer', 'nonValidId');
+        }
+    }
+
+    _checkSource(source) {
+        if (_.isUndefined(source)) {
+            throw new Error('layer', 'sourceRequired');
+        }
+        if (!(source instanceof SourceBase)) {
+            throw new Error('layer', 'nonValidSource');
+        }
+    }
+
+    _checkStyle(style) {
+        if (_.isUndefined(style)) {
+            throw new Error('layer', 'styleRequired');
+        }
+    }
+
+    /**
+     * [setSource description]
+     * @param {[type]} source [description]
+     * @api
+     */
     setSource(source) {
-        source._bindLayer(
+        this._checkSource(source);
+        source.bindLayer(
             dataframe => {
                 this._dataframes.push(dataframe);
                 this._mglIntegrator.renderer.addDataframe(dataframe);
@@ -36,14 +78,44 @@ export default class Layer {
         }
         this._source = source;
     }
+
+    /**
+     * [setStyle description]
+     * @param {[type]} style [description]
+     * @api
+     */
     setStyle(style) {
-        this._style.onChange(null);
-        style.onChange(this._styleChanged.bind(this));
+        this._checkStyle(style);
+        if (this._style) {
+            this._style.onChange(null);
+        }
         this._style = style;
+        this._style.onChange(this._styleChanged.bind(this));
+        // Force style changed event
         this._styleChanged();
     }
+
+    /**
+     * [addTo description]
+     * @param {[type]} map           [description]
+     * @param {[type]} beforeLayerID [description]
+     */
+    addTo(map, beforeLayerID) {
+        if (this._isMGLMap(map)) {
+            this._addToMGLMap(map, beforeLayerID);
+        }
+        else {
+            throw new Error('layer', 'nonValidMap');
+        }
+    }
+
+    _isMGLMap() {
+        // TODO: implement this
+        return true;
+    }
+
     _styleChanged() {
-        if (!this._mglIntegrator.invalidateMGLWebGLState) {
+        if (!(this._mglIntegrator && this._mglIntegrator.invalidateMGLWebGLState)) {
             return;
         }
         this._getData();
@@ -59,30 +131,29 @@ export default class Layer {
             }
         });
     }
+
     _getViewport() {
         if (this._mglIntegrator) {
             return this._mglIntegrator.renderer.getBounds();
         }
         throw new Error('?');
     }
+
     _getData() {
         if (!this._mglIntegrator.invalidateMGLWebGLState) {
             return;
         }
-        const r = this._source._getData(this._getViewport(), this._style.getMinimumNeededSchema());
+        const r = this._source.requestData(this._getViewport(), this._style.getMinimumNeededSchema());
         if (r) {
             this.metadataPromise = r;
             r.then(() => this._styleChanged());
         }
     }
-    addTo(map, beforeLayerID) {
-        // TODO check map type
-        this._addToMGLMap(map, beforeLayerID);
-    }
+
     _addToMGLMap(map, beforeLayerID) {
         map.on('load', () => {
             this._mglIntegrator = getMGLIntegrator(map);
-            this._mglIntegrator.addLayer(this._name, beforeLayerID, this._getData.bind(this), () => {
+            this._mglIntegrator.addLayer(this._id, beforeLayerID, this._getData.bind(this), () => {
                 this._dataframes.map(
                     dataframe => {
                         dataframe.setStyle(this._style);
